Migrate fourth fetch into utils
