+++ src/components/FavoriteContainer/FavoriteContainer.js
@@ -2,6 +2,7 @@ import React from 'react';
 import Nav from "../Nav/Nav";
 import './FavoriteContainer.scss';
 import ListingCard from '../ListingCard/ListingCard';
+import { fetchFavoriteListingData } from '../../utils';
 
 class FavoriteContainer extends React.Component {
   constructor() {
@@ -14,7 +15,7 @@ class FavoriteContainer extends React.Component {
 
   componentDidMount() {
     this.fetchSavedFavorites();
-    this.fetchListingData();
+    fetchFavoriteListingData().then(listings => this.setState({listings}));
   }
 
   componentDidUpdate() {
@@ -26,20 +27,6 @@ class FavoriteContainer extends React.Component {
        const favorites = JSON.parse(window.localStorage.getItem("listingFavorites"));
        this.setState({favorites: favorites});
      };
-  }
-
-  fetchListingData = () => {
-    const favorites = JSON.parse(window.localStorage.getItem("listingFavorites"));
-    const promises = favorites.map(listingID => {
-      return fetch('http://localhost:3001/api/v1/listings/' + listingID)
-      .then(res => res.json())
-      .then(data => {
-        return {listingID: listingID,
-          areaID: data.area_id,
-          name: data.name}
-      })
-    })
-    Promise.all(promises).then(listings => this.setState({listings}));
   }
 
   updateSavedFavorites = () => {
